Conform to the naming policy - from mx_AppsContainer to mx_AppsDrawer_appsContainer
The class name "mx_AppsContainer" is defined by const "appsContainer", and "appsContainer" is either equal to "drawer" (if widgetIsMaxmised is true) or included in it.

Since "drawer" is wrapped by a div, whose class name is defined by the const "classes", and one of the "classes" is "mx_AppsDrawer", it is sensible to regard "mx_AppsContainer" (mx_AppsDrawer_appsContainer) is defined by "mx_AppsDrawer".