Fix the weird gradient on the spaces bar
Really this mostly just needed a margin above the TAC button (which
the Add Space button did) to space it out from the, uh, spaces.
Having done that, I got annoyed by the fact that the gradients were
percentage-based and so the amount that got faded out depended on how
tall your element window was, so most of the changes here are to fix
the gradients to be fixed pixel offsets from the top & bottom.