Use latest LTS Node in CI
The README says we require the latest LTS version of Node, but actions/setup-node with ubuntu-latest appears to default to a version earlier than that. I'd like to make use of a feature (https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/unicodeSets) available only since Node 20, which happens to be the latest LTS. So, let's get CI on the right version.