Use LTS nodejs in github workflows
Currently our GH workflows all use nodejs 18, but that's not compatible with the
stated support range of matrix-js-sdk, which requires the latest LTS nodejs;
so, when we attermpt to `yarn install` on an older nodejs, we get an error.

Switch to LTS nodejs for all the places we set up nodejs.