Set entrypoints to use ./lib rather than ./src
Currently, we replace the entrypoints in package.json during the release
cycle. I think, historically, this was done to make element-web development
easier, but that doesn't actually use these entrypoints (instead it imports
from `src`).

Accordingly, I think the switcheroo is unnecessary; furthermore it causes a
whole bunch of confusion by making the development environment different from
the release environment, and it complicates our CI and release process.

In short, the switcheroo has to die.